moved to correct test file, and removed unneded reference
